Allow iteration over available file names
Your example program included an unimplemented `entries` method.  I
chose not to implement `entries` because iterating over it would require
decompressing any large files stored in the map.

Instead, I've provided a `file_names` API that provides just the names
of the files in the map, which the caller can look up (and potentially
decompress) if they wish.

I implemented a custom iterator wrapper to hide the implementation
details from the caller.